Safari-609.1.11.1/LayoutTests/svg/animations/svgnumber-animation-4-expected.txt

SVG 1.1 dynamic animation tests

Test for SVGNumber animation on SVG DOM properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stop1.offset.animVal is 0
PASS stop1.offset.baseVal is 0
PASS stop1.offset.animVal is 0.5
PASS stop1.offset.baseVal is 0
PASS stop1.offset.animVal is 1
PASS stop1.offset.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
